test: clean up stale skipped tests and dead code
- Delete 5 obsolete Bake/Rig gate tests from test_bake_rig_parity.py
  (replaced by sparse VMD/live-rig design)
- Delete stale test_animation_converter.py (AnimationConverter never existed)
- Remove unconditional skip from PMX/PMD multiple-file import tests;
  use fixture-count guard and add scene cleanup between imports
- Move _quat_angle_deg/WORLD_ROT_THRESHOLD_DEG to real_asset test
  (only remaining consumer)
- Remove dead code: _capture_bone_world_transforms, VERTEX_CASES,
  _import_bake, _import_rig

Integration: 105 PASS (22 skip), unit: 806 PASS (18 skip)
